Support long intertitles and fadeout effect
Intertitles can now span several lines -- with manual line breaks
in place of the "|" pipe character, or with automatic line breaks
where necessary -- and the whole text fits on the video. The width
of the text box is now 0.8 times the width of the video, this is
controlled by constant TEXT_WIDTH_FACTOR.

Added option --fadeout which accepts duration of the fadeout
in milliseconds.

Fixed option --intertitle-duration which was causing an index
out of range exception, because it was not being parsed as int.
The same goes for --intertitle-fontsize.